# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CodeSteppe API - Mongolian competitive coding platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hiring/challenges/{id}:
    get:
      operationId: getHiringChallenge
      tags: [hiring]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HiringChallengeDetail"
  /hiring/challenges/{id}/apply:
    post:
      operationId: applyToHiringChallenge
      tags: [hiring]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /hiring/challenges/{id}/leaderboard:
    get:
      operationId: getHiringLeaderboard
      tags: [hiring]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateEntry"
...
  /bosses/{slug}:
    get:
      operationId: getBoss
      tags: [bosses]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BossDetail"
  /bosses/{slug}/start:
    post:
      operationId: startBossFight
      tags: [bosses]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BossFightState"
...
  /boss-fights/{id}/attack:
    post:
      operationId: bossAttack
      tags: [bosses]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BossAttackInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BossAttackResult"
  /boss-fights/{id}/forfeit:
    post:
      operationId: forfeitBossFight
      tags: [bosses]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    ProblemSummary:
      type: object
      required: [id, slug, title, difficulty, tags, xpReward, eloReward, solvedCount]
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        difficulty: { type: string }
        tags:
          type: array
          items: { type: string }
        xpReward: { type: integer }
        eloReward: { type: integer }
        solvedCount: { type: integer }
        solved: { type: boolean }
    Problem:
      type: object
      required:
        [
          id,
          slug,
          title,
          difficulty,
          statement,
          inputDescription,
          outputDescription,
          constraints,
          examples,
          publicTestCases,
          tags,
          xpReward,
          eloReward,
          timeLimit,
          memoryLimit,
          starterCode,
        ]
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        difficulty: { type: string }
        statement: { type: string }
        inputDescription: { type: string }
        outputDescription: { type: string }
        constraints: { type: string }
        examples:
          type: array
          items:
            type: object
            required: [input, output]
            properties:
              input: { type: string }
              output: { type: string }
              explanation: { type: ["string", "null"] }
        publicTestCases:
          type: array
          items:
            type: object
            required: [input, expectedOutput]
            properties:
              input: { type: string }
              expectedOutput: { type: string }
        tags:
          type: array
          items: { type: string }
        xpReward: { type: integer }
        eloReward: { type: integer }
        timeLimit: { type: integer }
        memoryLimit: { type: integer }
        starterCode:
          type: object
          required: [javascript, python]
          properties:
            javascript: { type: string }
            python: { type: string }
...
    HiringChallengeDetail:
      type: object
      required: [id, title, description, companyUsername, companyDisplayName, problems, applicantCount, positions, applied, myStatus]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        companyUsername: { type: string }
        companyDisplayName: { type: string }
        problems:
          type: array
          items: { $ref: "#/components/schemas/ProblemSummary" }
        applicantCount: { type: integer }
        positions: { type: integer }
        closesAt:
          type: ["string", "null"]
          format: date-time
        applied: { type: boolean }
        myStatus: { type: string }
        myScore: { type: ["integer", "null"] }
        mySolved: { type: ["integer", "null"] }
    CandidateEntry:
      type: object
      required: [position, username, displayName, score, solvedCount, status]
      properties:
        position: { type: integer }
        username: { type: string }
        displayName: { type: string }
        avatarSeed: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        score: { type: integer }
        solvedCount: { type: integer }
        status: { type: string }
        eloRating: { type: integer }
        rank: { type: string }
...
    BossDetail:
      type: object
      required: [id, slug, name, title, description, difficulty, maxHp, rewardXp, rewardCoins, problems, artColor, icon, defeated]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        title: { type: string }
        description: { type: string }
        difficulty: { type: string }
        maxHp: { type: integer }
        rewardXp: { type: integer }
        rewardCoins: { type: integer }
        rewardTitle: { type: ["string", "null"] }
        problems:
          type: array
          items: { $ref: "#/components/schemas/ProblemSummary" }
        artColor: { type: string }
        icon: { type: string }
        defeated: { type: boolean }
    BossFightState:
      type: object
      required: [id, bossId, bossSlug, bossName, maxHp, bossHp, playerHp, combo, state, currentProblem, totalProblems, currentProblemIdx]
      properties:
        id: { type: integer }
        bossId: { type: integer }
        bossSlug: { type: string }
        bossName: { type: string }
        title: { type: string }
        artColor: { type: string }
        icon: { type: string }
        maxHp: { type: integer }
        bossHp: { type: integer }
        playerHp: { type: integer }
        combo: { type: integer }
        state:
          type: string
          enum: [active, victory, defeat]
        currentProblem:
          oneOf:
            - $ref: "#/components/schemas/Problem"
            - type: "null"
        currentProblemIdx: { type: integer }
        totalProblems: { type: integer }
        result: { type: ["string", "null"] }
        rewardXp: { type: ["integer", "null"] }
        rewardCoins: { type: ["integer", "null"] }
        rewardTitle: { type: ["string", "null"] }
    BossAttackInput:
      type: object
      required: [language, code]
      properties:
        language:
          type: string
          enum: [javascript, typescript, python, cpp]
        code: { type: string }
    BossAttackResult:
      type: object
      required: [passed, totalCount, passedCount, damageDealt, playerDamage, comboMultiplier, message, fight]
      properties:
        passed: { type: boolean }
        passedCount: { type: integer }
        totalCount: { type: integer }
        damageDealt: { type: integer }
        playerDamage: { type: integer }
        comboMultiplier: { type: number }
        message: { type: string }
        fight: { $ref: "#/components/schemas/BossFightState" }
...